+++ src/index.html
@@ -13,8 +13,8 @@
             <button class="btn ref-btn" id="refresh-btn" type="button"></button>
         </div>
         <div class="card">
-            <input type="text" class="inp" placeholder="Add to your list...">
-            <button class="btn insert-btn" id="enter" type="button"></button>
+            <input id="insert" type="text" class="inp" placeholder="Add to your list...">
+            <button class="btn insert-btn" id="enterBtn" type="button"></button>
         </div>
         <div id="list"></div>
         <div class="card card-clear">
