+++ src/test.html
@@ -1,0 +1,75 @@
+let myForm = document.getElementById('myForm');
+let insert = document.getElementById('insert');
+let enterBtn = document.getElementById('enterBtn');
+let toDoList = JSON.parse(localStorage.getItem('ToDo-List')) || [];
+
+
+enterBtn.addEventListener('click', () => {
+    
+    addnew(insert.value);
+    insert.value = '';
+    
+});
+
+const addnew = (newTask) => {
+    if (newTask === '') return;
+    let newToDo = {
+        description: newTask,
+        completed: false,
+        index: toDoList.length,
+    };
+    const newToDoList = [...toDoList,newToDo];
+    localStorage.setItem('ToDo-List', JSON.stringify(newToDoList));
+    
+    
+};
+
+const displayItems = () => {
+    if (toDoList === null) return;
+    const listCard = document.getElementById('list');
+    listCard.innerHTML = '';
+    toDoList.forEach((arr) => {
+        const toDo = `
+        <div class="card-task">
+            <input type="checkbox" id="checkbox" data-id="${arr.index}">
+            <input type="text" id="editInp" class="inp" value="${arr.description}" data-id="${arr.index}">
+            <button id="remove" class="btn btn-drag" type="button" data-remove="${arr.index}"></button>
+        <div>
+        `;
+        listCard.insertAdjacentHTML('beforeend', toDo);
+    });
+    
+    }
+    displayItems();
+    
+    // Editing Task
+    const editInput = document.querySelectorAll('#editInp');
+    editInput.forEach((item) => {
+        item.addEventListener('keyup', () => {
+            let itemId = item.getAttribute('data-id');
+            let itemValue = item.value;
+            let existingTask = toDoList.find((t) => t.index == itemId);
+            existingTask.description = itemValue;
+            localStorage.setItem('ToDo-List', JSON.stringify(toDoList));
+        })
+    } )
+ 
+    // Deleting Task
+    const remove = (taskToDelete) => {
+        newList = toDoList.filter((task, i) => i === taskToDelete);
+        console.log(newList)
+        localStorage.setItem('ToDo-List', JSON.stringify(newList));
+    }
+
+    const removeBtn = document.querySelectorAll('#remove');
+    const removeTask = document.querySelectorAll('#checkbox');
+    removeBtn.forEach((e) => {
+        e.addEventListener('click', () => {
+            let taskToDelete = e.getAttribute('data-remove');
+            remove(taskToDelete);
+            
+        })
+        
+    })
+
+
